One byte is enough for ChronologyLogLevel in binary form

+++ chronology-log4j-1/src/main/java/com/higherfrequencytrading/chronology/log4j1/BinaryChronicleAppender.java
@@ -47,7 +47,7 @@ public abstract class BinaryChronicleAppender extends AbstractChronicleAppender 
             appender.writeByte(Chronology.VERSION);
             appender.writeByte(Chronology.TYPE_LOG4J_1);
             appender.writeLong(event.getTimeStamp());
-            appender.writeInt(toIntChronologyLogLevel(event.getLevel()));
+            appender.writeByte(toIntChronologyLogLevel(event.getLevel()));
             appender.writeUTF(event.getThreadName());
             appender.writeUTF(event.getLoggerName());
             appender.writeUTF(event.getMessage().toString());
